requirements.txt
hikari
hikari-lightbulb
python-dotenv

# Used by Natural Readers
httpx
httpx-auth